[IPNode] Move logdir creation to IPNode
Based on the review suggestion from @oliviertilmans, logdir creation is
a desirable feature for all sorts of daemons. To generalize the logdir
creation, move the directory creation to `IPNode`. Add a `logdir`
property to the basic `Daemon` class which reads the logdir from the
`logfile` option. The `Openr` class must overwrite this property since
it does not use the `logfile`, but a `log_dir` option.

Directories are now created using `pexec` instead of `os.mkdirs`.
Creating the directory with pexec has an additional neat feature. When
the nodes `logdir` is private, logdirs are created in the directory path
of the node's private folder.